moved EmSize to bevy_text/src/text.rs next to RemSize. Feels neater.
